+++ apps/mail_composer/mail_composer_app.rb
@@ -18,7 +18,7 @@ class MailComposer < MonadicApp
 
   def settings
     {
-      "model": "gpt-4-turbo-preview",
+      "model": "gpt-3.5-turbo-1106",
       "temperature": 0.3,
       "top_p": 0.0,
       "max_tokens": 1000,