+++ apps/math_tutor/math_tutor_app.rb
@@ -19,7 +19,7 @@ class MathTutor < MonadicApp
 
   def settings
     {
-      "model": "gpt-4-turbo-preview",
+      "model": "gpt-3.5-turbo-1106",
       "temperature": 0.0,
       "top_p": 0.0,
       "max_tokens": 2000,